requirements.txt
Flask==3.1.3
Flask-CORS==6.0.2
Flask-SQLAlchemy==3.1.1
fpdf2==2.8.6
